Add files via upload
This guide can be followed to create an open-source, low-cost system for continuous nitrate monitoring in soil and open water. If tube lengths or diameters differ, then the code might have to be changed. It is recommended to calibrate the vacuum suction timings, as well as the spectrophotometer, prior to leaving the system automated. For more information on how to create the spectrophotometer portion of the system: https://www.sciencedirect.com/science/article/pii/S246806722030016X